backend/requirements.txt
ipython
fastapi[all]
python-socketio
python-dotenv
httpx
celery
uvicorn
pydantic
redis
